refactor: got rid of Internal files
Just added complexity and not a library
