+++ package.yaml
@@ -37,14 +37,10 @@ library:
   - Taskell.Data.Date
   - Taskell.Data.Date.RelativeParser
   - Taskell.Data.List
-  - Taskell.Data.List.Internal
   - Taskell.Data.Lists
-  - Taskell.Data.Lists.Internal
   - Taskell.Data.Seq
   - Taskell.Data.Subtask
-  - Taskell.Data.Subtask.Internal
   - Taskell.Data.Task
-  - Taskell.Data.Task.Internal
   - Taskell.Events.Actions.Types
   - Taskell.Events.State.History
   - Taskell.Events.State.Types
